Turn the game server into a proper web server
Using snap intead of warp/wai simply because this project is already
building on top of snap.

Also puts the server state handling into a different module than the
main function. This now takes the usual server arguments.

This will require changes to the https reverse proxy.
