+++ codeworld-game-server/codeworld-game-server.cabal
@@ -14,6 +14,9 @@ executable codeworld-game-server
   build-depends:       base >=4.8 && <4.9,
                        text,
                        websockets == 0.9.*,
+                       websockets-snap == 0.10.*,
+                       snap-core == 1.0.*,
+                       snap-server == 1.0.*,
                        bytestring,
                        random,
                        unordered-containers,
@@ -21,3 +24,4 @@ executable codeworld-game-server
                        codeworld-game-api
   hs-source-dirs:      src
   default-language:    Haskell2010
+  ghc-options:         -threaded -rtsopts "-with-rtsopts=-N"
